% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_train.R
\name{accuracy.meas}
\alias{accuracy.meas}
\title{Calculate classification accuracy statistics.}
\usage{
accuracy.meas(response, predicted, threshold = 0.5)
}
\arguments{
\item{response}{Vector of actual wins.}

\item{predicted}{Vector of predicted win probabilities.}

\item{threshold}{Probability level that determines whether the
prediction is a win or a loss.}
}
\value{
Object of class \code{accuracy.meas} containing precision,
recall, and F-score statistics.
}
\description{
Calculate classification accuracy statistics.
}

