% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winprob.R
\name{best_kicking_option}
\alias{best_kicking_option}
\title{Use the expected win probabilities to determine best kicking option}
\usage{
best_kicking_option(probs, wp_ev_goforit)
}
\arguments{
\item{probs}{}

\item{wp_ev_goforit}{}
}
\description{
Use the expected win probabilities to determine best kicking option
}

