% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winprob.R
\name{breakeven}
\alias{breakeven}
\title{Calculates the breakeven point for making the decision.}
\usage{
breakeven(probs)
}
\arguments{
\item{probs}{}
}
\description{
The breakeven is the point at which a coach should be indifferent
between two options. We compare the expected win probability
of going for it on 4th down to the next best kicking option
and determine what the probability of converting the 4th down
needs to be in order to make the coach indifferent to going for it
or kicking.
}

