% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winprob.R
\name{calc_prob_success}
\alias{calc_prob_success}
\title{Use historical first down rates. When inside the opponent's 10,
use dwn, ytg, yfog specific rates. Otherwise, use binned yfog where
field is broken into 10 segments}
\usage{
calc_prob_success(situation, data)
}
\arguments{
\item{situation}{}

\item{data}{}
}
\description{
Use historical first down rates. When inside the opponent's 10,
use dwn, ytg, yfog specific rates. Otherwise, use binned yfog where
field is broken into 10 segments
}

