% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winprob.R
\name{calculate_features}
\alias{calculate_features}
\title{Generate features needed for the win probability model that are
not contained in the general game state information passed via API.}
\usage{
calculate_features(situation, data)
}
\arguments{
\item{situation}{List}

\item{data}{The same list, with new keys and values}
}
\description{
Generate features needed for the win probability model that are
not contained in the general game state information passed via API.
}

