% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{calculate_prob_poss}
\alias{calculate_prob_poss}
\title{Determine the starting point for the final possession in each
non-overtime, regular season game to use as a proxy for the probability
that the team will have another possession in the game during the 4th quarter.}
\usage{
calculate_prob_poss(drive_fname, out_name, games)
}
\arguments{
\item{drive_fname}{Filename of where DRIVE table is stored}

\item{out_name}{Filename of where to write output to}

\item{games}{Data frame generated by \code{load_games}}
}
\value{
data_frame
}
\description{
Used to weight the win probabilities in the 4th quarter.
}

