% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_train.R
\name{calibration_plot}
\alias{calibration_plot}
\title{Produces a calibration plot for the win probability model.}
\usage{
calibration_plot(preds, truth, fg = FALSE)
}
\arguments{
\item{preds}{Vector of predicted win probabilities.}

\item{truth}{Vector of actual wins.}
}
\value{
\code{ggplot2}
}
\description{
Splits the predictions into percentiles and calculates the percentage
of predictions per percentile that were wins. A perfectly calibrated model
means that plays with a win probability of n% win about n% of the time.
}

