% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plays.R
\name{change_poss}
\alias{change_poss}
\title{Handles situation updating for all plays that involve
a change of possession, including punts, field goals,
missed field goals, touchdowns, turnover on downs.}
\usage{
change_poss(situation, play_type, features, ...)
}
\arguments{
\item{situation}{List}

\item{play_type}{function}

\item{features}{List of existing features}

\item{...}{Additional parameters}
}
\value{
List
}
\description{
Handles situation updating for all plays that involve
a change of possession, including punts, field goals,
missed field goals, touchdowns, turnover on downs.
}

