% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{code_fourth_downs}
\alias{code_fourth_downs}
\title{Parse all fourth downs and determine if teams intended to go for it,
punt, or attempt a field goal. If intent is not clear, do not include
the play.}
\usage{
code_fourth_downs(df)
}
\arguments{
\item{df}{Play-by-play data frame}
}
\value{
data_frame
}
\description{
Parse all fourth downs and determine if teams intended to go for it,
punt, or attempt a field goal. If intent is not clear, do not include
the play.
}

