% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{data_prep}
\alias{data_prep}
\title{Prepare Armchair Analysis data for analysis.}
\usage{
data_prep(pbp_data_location)
}
\arguments{
\item{pbp_data_location}{Directory where Armchair Analysis csv
files are stored.}
}
\value{
None
}
\description{
Prepare Armchair Analysis data for analysis.
}

