% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{dp_kneel_down}
\alias{dp_kneel_down}
\title{Code a situation a 1 if the offense can kneel to end the game based on time
remaining, defensive timeouts remaining, down, and score difference.}
\usage{
dp_kneel_down(df)
}
\arguments{
\item{df}{Joined play-by-play data.}
}
\value{
data_frame
}
\description{
Code a situation a 1 if the offense can kneel to end the game based on time
remaining, defensive timeouts remaining, down, and score difference.
}

