% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winprob.R
\name{expected_win_prob}
\alias{expected_win_prob}
\title{Expected value of win probability, factoring in p(success).}
\usage{
expected_win_prob(pos_prob, pos_win_prob, neg_win_prob)
}
\arguments{
\item{pos_prob}{}

\item{pos_win_prob}{}

\item{neg_win_prob}{}
}
\description{
Expected value of win probability, factoring in p(success).
}

