% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winprob.R
\name{expected_wp_fg}
\alias{expected_wp_fg}
\title{Expected WP from kicking, factoring in p(FG made).}
\usage{
expected_wp_fg(situation, probs, data)
}
\arguments{
\item{situation}{}

\item{probs}{}

\item{data}{}
}
\description{
Expected WP from kicking, factoring in p(FG made).
}

