% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{fg_success_rate}
\alias{fg_success_rate}
\title{Historical field goal success rates by field position.}
\usage{
fg_success_rate(fg_data_fname, out_fname, min_pid = 473957)
}
\arguments{
\item{fg_data_fname}{Filename of Armchair Analysis GAME table}

\item{out_fname}{Filename of where to write output to}

\item{min_pid}{Starting point in play-by-play data of which kicks to include.
Defaults to start of 2011 season}
}
\value{
data_frame
}
\description{
By default, uses only attempts from >= 2011 season to reflect
more improved kicker performance.
}
\details{
Returns and writes results to a CSV.

NOTE: These are somewhat sparse and irregular at longer FG ranges.
This is because kickers who attempt long FGs are not selected at
random -- they are either in situations which require a long FG
attempt or are kickers with a known long range. The NYT model
uses a logistic regression kicking model developed by Josh Katz
to smooth out these rates.
}

