% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{first_down_rates}
\alias{first_down_rates}
\title{Find the mean 1st down success rate at a given point in the field.}
\usage{
first_down_rates(df_plays, yfog)
}
\arguments{
\item{df_plays}{Play by play data frame}

\item{yfog}{String, must be 'yfog' or 'yfog_bin'. If yfog, use the actual yards
from own goal. If yfog_bin, use the decile of the field instead.}
}
\value{
data_frame
}
\description{
Find the mean 1st down success rate at a given point in the field.
}

