% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winprob.R
\name{generate_decision}
\alias{generate_decision}
\title{Decide on optimal play based on game states and their associated
win probabilities.}
\usage{
generate_decision(situation, data, probs)
}
\arguments{
\item{situation}{}

\item{data}{}

\item{probs}{}
}
\description{
Note the currently 'best play' is based purely
on the outcome with the highest expected win probability. This
does not account for uncertainty of these estimates.
For example, the win probabilty added by a certain play may be
very small (0.0001), but that may be the 'best play.'
}

