% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winprob.R
\name{generate_response}
\alias{generate_response}
\title{Parent function called by the bot to make decisions on 4th downs.}
\usage{
generate_response(situation, data, model)
}
\arguments{
\item{situation}{List}

\item{data}{List, contains historical data frames}

\item{model}{Logistic regression model, results of \code{logit}}
}
\value{
List
}
\description{
Parent function called by the bot to make decisions on 4th downs.
}

