% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winprob.R
\name{generate_win_probabilities}
\alias{generate_win_probabilities}
\title{For each of the possible scenarios, estimate the win probability
for that game state.}
\usage{
generate_win_probabilities(situation, scenarios, model, data)
}
\arguments{
\item{situation}{}

\item{scenarios}{}

\item{model}{}

\item{data}{}
}
\description{
For each of the possible scenarios, estimate the win probability
for that game state.
}

