% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winprob.R
\name{get_historical_decision}
\alias{get_historical_decision}
\title{Compare current game situation to historically similar situations.}
\usage{
get_historical_decision(situation, data, decision)
}
\arguments{
\item{situation}{}

\item{data}{}

\item{decision}{}
}
\description{
Currently uses score difference and field position to provide
rough guides to what coaches have done in the past.
}

