% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{group_coaches_decisions}
\alias{group_coaches_decisions}
\title{Group 4th down decisions by score difference and field
position to get coarse historical comparisons.}
\usage{
group_coaches_decisions(fourths, out_fname)
}
\arguments{
\item{fourths}{Joined play-by-play and game data}
}
\value{
data_frame
}
\description{
Writes these to a CSV and returns them.
}

