% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{join_df_first_down_rates}
\alias{join_df_first_down_rates}
\title{Join the computed first down rates with the play by play data.}
\usage{
join_df_first_down_rates(df, fd_open_field, fd_inside_10)
}
\arguments{
\item{df}{Joined play-by-play data.}

\item{fd_open_field}{Results of \code{first_down_rates} with yfog set to 'yfog_bin'.}

\item{fd_inside_10}{Results of \code{first_down_rates} with yfog set to 'yfog'.}
}
\value{
data_frame
}
\description{
Join the computed first down rates with the play by play data.
}

