% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plays.R
\name{kneel_down}
\alias{kneel_down}
\title{kneel_down}
\usage{
kneel_down(score_diff, timd, secs_left, dwn)
}
\arguments{
\item{score_diff}{}

\item{timd}{}

\item{secs_left}{}

\item{dwn}{}
}
\value{
0/1
}
\description{
kneel_down
}

