% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bot.R
\name{load_data}
\alias{load_data}
\title{Load previously compiled data and estimated win and field goal models.}
\usage{
load_data()
}
\description{
Load previously compiled data and estimated win and field goal models.
}

