% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{load_games}
\alias{load_games}
\title{Load data containing results of each game and return a DataFrame.}
\usage{
load_games(game_data_fname, remove_ties = FALSE)
}
\arguments{
\item{game_data_fname}{Filename of Armchair Analysis GAME table}

\item{remove_ties}{Optional}
}
\value{
data_frame
}
\description{
Load data containing results of each game and return a DataFrame.
}

