% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{load_pbp}
\alias{load_pbp}
\title{Load the play by play data and return a data frame.}
\usage{
load_pbp(pbp_data_fname, games, remove_knees = FALSE)
}
\arguments{
\item{pbp_data_fname}{Location of play by play data}

\item{games}{Data frame, game-level data frame created by \code{load_games}}

\item{remove_knees}{Optional}
}
\value{
data_frame
}
\description{
Load the play by play data and return a data frame.
}

