% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plays.R
\name{missed_field_goal}
\alias{missed_field_goal}
\title{Opponent takes over from the spot of the kick.}
\usage{
missed_field_goal(situation, new_situation)
}
\arguments{
\item{situation}{}

\item{new_situation}{}
}
\description{
Opponent takes over from the spot of the kick.
}

