% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_train.R
\name{model_train}
\alias{model_train}
\title{Estimate win probability model using Armchair Analysis play-by-play data.}
\usage{
model_train(plot = FALSE)
}
\arguments{
\item{plot}{Set to 'TRUE' if you wish to view the calibration plots
and ROC curves for the model.}
}
\value{
None
}
\description{
Estimate win probability model using Armchair Analysis play-by-play data.
}

