% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{nyt_fg_model}
\alias{nyt_fg_model}
\title{Sub in simple logit for field goal success rates.}
\usage{
nyt_fg_model(fname, outname)
}
\arguments{
\item{fname}{Filename of NYT model output}

\item{outname}{Filename of where to write output to}
}
\value{
data_frame
}
\description{
Sub in simple logit for field goal success rates.
}

