% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winprob.R
\name{plot_decision}
\alias{plot_decision}
\title{Create a plot of the decision state to go for it, attempt a field goal,
or punt.}
\usage{
plot_decision(results)
}
\arguments{
\item{results}{Result of \code{generate_decision}.}
}
\description{
Create a plot of the decision state to go for it, attempt a field goal,
or punt.
}

