% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_train.R
\name{plot_roc}
\alias{plot_roc}
\title{ROC Plot}
\usage{
plot_roc(roc_data, roc_auc)
}
\arguments{
\item{roc_data}{Data frame containing ROCR \code{performance} results.}

\item{roc_auc}{AUC value.}
}
\value{
Plot of ROC curve.
}
\description{
ROC Plot
}

