% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plays.R
\name{punt}
\alias{punt}
\title{Use the average net punt distance (punt distance - return yards).}
\usage{
punt(situation, new_situation, ...)
}
\arguments{
\item{situation}{}

\item{new_situation}{}

\item{...}{}
}
\description{
Not all situations have historical data, especially very
close to opponent's end zone. Use a net punt distance of
5 yards here.
}

