% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{punt_averages}
\alias{punt_averages}
\title{Group punts by kicking field position to get average return distance.
Currently does not incorporate the possibility of a muffed punt
or punt returned for a TD.}
\usage{
punt_averages(punt_data_fname, out_fname, joined)
}
\arguments{
\item{punt_data_fname}{Filename of where PUNT table is stored}

\item{out_fname}{Filename of where to write output to}

\item{joined}{Joined play-by-play and game data}
}
\value{
data_frame
}
\description{
Group punts by kicking field position to get average return distance.
Currently does not incorporate the possibility of a muffed punt
or punt returned for a TD.
}

