% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plays.R
\name{qtr}
\alias{qtr}
\title{Given the seconds left in the game, determine the current quarter.}
\usage{
qtr(secs_left)
}
\arguments{
\item{secs_left}{}
}
\description{
Given the seconds left in the game, determine the current quarter.
}

