% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winprob.R
\name{random_play}
\alias{random_play}
\title{Generate a random play with plausible values for debugging purposes.}
\usage{
random_play(data)
}
\arguments{
\item{data}{}
}
\description{
Generate a random play with plausible values for debugging purposes.
}

