% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bot.R
\name{run_bot}
\alias{run_bot}
\title{Calculate the expected win probability for a given game state.}
\usage{
run_bot(data, win_model, fg_model, situation = NULL)
}
\arguments{
\item{data}{}

\item{win_model}{}

\item{fg_model}{}

\item{situation}{}
}
\description{
Calculate the expected win probability for a given game state.
}

