% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winprob.R
\name{simulate_scenarios}
\alias{simulate_scenarios}
\title{Simulate game state after each possible outcome.
Possible scenarios are: touchdown, first down, turnover on downs,
field goal attempt (success or failure), and punt.}
\usage{
simulate_scenarios(situation, data)
}
\arguments{
\item{situation}{}

\item{data}{}
}
\value{
scenarios
}
\description{
Simulate game state after each possible outcome.
Possible scenarios are: touchdown, first down, turnover on downs,
field goal attempt (success or failure), and punt.
}

