% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{switch_offense}
\alias{switch_offense}
\title{Swap game state columns for offense & defense dependent variables.
The play by play data has some statistics on punts and kickoffs in terms
of the receiving team. Switch these to reflect the game state for
the kicking team.}
\usage{
switch_offense(df)
}
\arguments{
\item{df}{Play-by-play data frame}
}
\value{
data_frame
}
\description{
Swap game state columns for offense & defense dependent variables.
The play by play data has some statistics on punts and kickoffs in terms
of the receiving team. Switch these to reflect the game state for
the kicking team.
}

